Move rules to its own folder
