+++ src/rule/rules/isCalculatedAreaTooBig.ts
@@ -1,7 +1,7 @@
 import {
     MAX_CLUSTER_SIZE
 } from '../../../SETTINGS';
-import { Rule } from '../rule';
+import { Rule } from '../index';
 
 export const isCalculatedAreaTooBig = () => new Rule(
     3,
